Explicitly cast untyped num in HxArray

+++ test/unit_testing/tests/RestArgs/intended/include/HxArray.h
@@ -150,7 +150,7 @@ public:
 		int tempResult;
 		
 		if(it != a->end()) {
-			tempResult = it - a->begin();
+			tempResult = ((int)(it - a->begin()));
 		} else {
 			tempResult = -1;
 		};
